PaloAlto Manager for user management
The script assumes you want to create "Superusers". If you need specific roles (e.g., read-only), the XML payload in update_user_password needs to be modified.

Self-Deletion: Be careful not to delete the user you are currently using to run the script, or the commit might fail (or you'll lock yourself out).